fixed assembly name and removed runcoff.cs
