+++ CoffLoader/src/CoffParser.cs
@@ -4,7 +4,7 @@ using System.Diagnostics;
 using System.Collections.Generic;
 using System.Runtime.InteropServices;
 
-namespace COFFLoader
+namespace CoffLoader
 {
     unsafe class CoffParser
     {
